1. created new dir for custom hooks 2. created new custom hook for clearning data about current quiz
