+++ src/components/pages/MainQuizPage/EndQuizButton/EndQuizButton.tsx
@@ -1,14 +1,16 @@
 import { useState } from "react";
-import { useNavigate } from "react-router-dom";
 import { StyledButton } from "../../../spreadedComponents/StyledButton/StyledButton";
 import { ModalPage } from "../../ModalPage/ModalPage";
-import { START } from "../../../Router/routes";
+import { useClearCurrentQuizData } from "../../../../hooks/useClearCurrentQuizData";
 
 const EndQuizButton = () => {
     const [showModal, setShowModal] = useState(false);
-    const navigate = useNavigate();
     const onCloseModalHandler = () => setShowModal(!showModal);
-    const onNavigateToResultHandler = () => navigate(START);
+    const clearCurrentQuizData = useClearCurrentQuizData();
+
+    const onConfirmHandler = () => {
+        clearCurrentQuizData();
+    };
 
     return (
         <>
@@ -16,7 +18,7 @@ const EndQuizButton = () => {
             <ModalPage
                 isOpen={showModal}
                 setShowModal={onCloseModalHandler}
-                onConfirm={onNavigateToResultHandler}
+                onConfirm={onConfirmHandler}
             >
                 <div className="title">
                     <h1>Are you sure you want to finish quiz?</h1>
