+++ src/redux/store/store.ts
@@ -16,7 +16,7 @@ const newStore = configureStore({
     }
 });
 
-export type RootState = ReturnType<typeof newStore.getState>
-export type AppDispatch = typeof newStore.dispatch
+export type RootState = ReturnType<typeof newStore.getState>;
+export type AppDispatch = typeof newStore.dispatch;
 
 export { newStore };
